api/services/api/requirements.txt
Flask==2.2.2
flask_swagger
gunicorn==20.1.0
cbor2
requests
elasticsearch
urllib3
python-dotenv
jsonschema
PyPDF2
python-dateutil
xmltodict
